requirements.txt
streamlit
streamlit-webrtc
av
opencv-python
tensorflow==2.16.1  # or tensorflow-cpu==2.16.1
Pillow
numpy
